Hypothetical mean

+++ app/src/main/java/com/sharpdroid/registroelettronico/utils/LayoutHelper.java
@@ -7,11 +7,11 @@ public class LayoutHelper {
     public static final int MATCH_PARENT = -1;
     public static final int WRAP_CONTENT = -2;
 
-    private static int getSize(float size) {
-        return (int) (size < 0 ? size : Metodi.dp(size));
+    private static int getSize(int size) {
+        return (size < 0 ? size : Metodi.dp(size));
     }
 
-    public static FrameLayout.LayoutParams createFrame(int width, float height, int gravity, float leftMargin, float topMargin, float rightMargin, float bottomMargin) {
+    public static FrameLayout.LayoutParams createFrame(int width, int height, int gravity, float leftMargin, float topMargin, float rightMargin, float bottomMargin) {
         FrameLayout.LayoutParams layoutParams = new FrameLayout.LayoutParams(getSize(width), getSize(height), gravity);
         layoutParams.setMargins(Metodi.dp(leftMargin), Metodi.dp(topMargin), Metodi.dp(rightMargin), Metodi.dp(bottomMargin));
         return layoutParams;
