+++ app/src/main/java/com/sharpdroid/registroelettronico/views/cells/ValueDetailsCheckboxCell.kt
@@ -33,7 +33,7 @@ class ValueDetailsCheckboxCell(context: Context) : FrameLayout(context) {
         dividerPaint.color = -0x262627
 
         checkBox = CheckBox(context)
-        addView(checkBox, LayoutHelper.createFrame(LayoutHelper.WRAP_CONTENT, LayoutHelper.WRAP_CONTENT.toFloat(), Gravity.START or Gravity.CENTER_VERTICAL, 8f, 0f, 16f, 0f))
+        addView(checkBox, LayoutHelper.createFrame(LayoutHelper.WRAP_CONTENT, LayoutHelper.WRAP_CONTENT, Gravity.START or Gravity.CENTER_VERTICAL, 8f, 0f, 16f, 0f))
 
         textView = TextView(context)
         textView.setTextColor(Color.BLACK)
@@ -43,7 +43,7 @@ class ValueDetailsCheckboxCell(context: Context) : FrameLayout(context) {
         textView.setSingleLine(true)
         textView.ellipsize = TextUtils.TruncateAt.END
         textView.gravity = Gravity.START or Gravity.CENTER_VERTICAL
-        addView(textView, LayoutHelper.createFrame(LayoutHelper.WRAP_CONTENT, LayoutHelper.WRAP_CONTENT.toFloat(), Gravity.START or Gravity.TOP, 57f, 10f, 17f, 0f))
+        addView(textView, LayoutHelper.createFrame(LayoutHelper.WRAP_CONTENT, LayoutHelper.WRAP_CONTENT, Gravity.START or Gravity.TOP, 57f, 10f, 17f, 0f))
 
         valueTextView = TextView(context)
         valueTextView.setTextColor(-0x757576)
@@ -53,7 +53,7 @@ class ValueDetailsCheckboxCell(context: Context) : FrameLayout(context) {
         valueTextView.maxLines = 1
         valueTextView.setSingleLine(true)
         valueTextView.setPadding(0, 0, 0, 0)
-        addView(valueTextView, LayoutHelper.createFrame(LayoutHelper.WRAP_CONTENT, LayoutHelper.WRAP_CONTENT.toFloat(), Gravity.START or Gravity.TOP, 57f, 35f, 17f, 0f))
+        addView(valueTextView, LayoutHelper.createFrame(LayoutHelper.WRAP_CONTENT, LayoutHelper.WRAP_CONTENT, Gravity.START or Gravity.TOP, 57f, 35f, 17f, 0f))
     }
 
     fun setTextAndValue(text: String, value: String, divider: Boolean) {
